Integrated the encounter mechanic system and "strategy" pattern system of switching between movement systems instead of just having a big gross if check system. Still need to do some work on the encounter mechanic but the one that omar had is working.
