+++ SoteriaGame/Assets/Scripts/Controllers/StandardMovementController.cs.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 09dd6d3416d679246a4881fc8ee050c1
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData:
